+++ lib/features/working_directory/presentation/ui/working_directory_item.dart
@@ -1,17 +1,17 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:open_git/features/working_directory/presentation/bloc/working_directory_bloc.dart';
 import 'package:open_git/shared/domain/entities/git_file_entity.dart';
 import 'package:open_git/shared/presentation/widgets/gaps.dart';
 
 class WorkingDirectoryItem extends StatelessWidget {
   final GitFileEntity file;
   final Function(GitFileEntity file) onSelected;
-  final ValueChanged<bool?> onCheckboxToggled;
 
   const WorkingDirectoryItem({
     super.key,
     required this.file,
     required this.onSelected,
-    required this.onCheckboxToggled,
   });
 
   @override
@@ -24,8 +24,10 @@ class WorkingDirectoryItem extends StatelessWidget {
         children: [
           Checkbox(
             value: file.staged,
-            onChanged: (value) {
-              onCheckboxToggled(value ?? false);
+            onChanged: (_) {
+              context.read<WorkingDirectoryBloc>().add(
+                ToggleFileStaging(file: file, stage: !file.staged),
+              );
             },
           ),
           Icon(